Generate desert (#1)
* add shader wrapper (need to be integrated to the project)

* fix compilation

* Finalize shader split in files

* start objParser

* try to link OBJ_Loader lib

* (WIP) load map

* fix OBJ_Loader warnings

* Load and display obj

* remove useless stuff

* removing OBJ_Loader + removing sphere generation code

+++ include/renderer.hpp
@@ -1,6 +1,9 @@
 #ifndef RENDERER_HPP
 #define RENDERER_HPP
 
+#include "shader/ShaderLinker.hpp"
+#include "shader/Shader.hpp"
+
 BEGIN_VISUALIZER_NAMESPACE
 
 class Camera;
@@ -28,7 +31,7 @@ public:
     void UpdateCamera();
 
 private:
-    uint16_t m_IndexCount;
+    int m_IndexCount;
     GLuint m_VAO, m_VBO, m_IBO, m_UBO, m_ShaderProgram;
 
     glm::mat4* m_UBOData = nullptr;
